Autorización para el apartado de Proyectos

+++ app/Http/Controllers/API/ProyectoController.php
@@ -6,14 +6,24 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\ProyectoResource;
 use App\Models\Proyecto;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 use PhpParser\Node\Stmt\TryCatch;
 
-class ProyectoController extends Controller
+class ProyectoController extends Controller implements HasMiddleware
 {
     public $modelclass = Proyecto::class;
     /**
      * Display a listing of the resource.
      */
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show']),
+        ];
+    }
 
     public function index(Request $request)
     {
@@ -31,8 +41,11 @@ class ProyectoController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Proyecto::class);
         $proyecto = json_decode($request->getContent(), true);
-
+        if (!$request->user()->esAdmin()) {
+            $proyecto['docente_id'] = $request->user()->id;
+        }
         $proyecto = Proyecto::create($proyecto);
 
         return new ProyectoResource($proyecto);
@@ -43,7 +56,7 @@ class ProyectoController extends Controller
      */
     public function show(Proyecto $proyecto)
     {
-            return new ProyectoResource($proyecto);
+        return new ProyectoResource($proyecto);
     }
 
     /**
@@ -51,6 +64,8 @@ class ProyectoController extends Controller
      */
     public function update(Request $request, Proyecto $proyecto)
     {
+        Gate::authorize('update', $proyecto);
+
         $proyectoData = json_decode($request->getContent(), true);
         $proyecto->update($proyectoData);
 
@@ -62,6 +77,7 @@ class ProyectoController extends Controller
      */
     public function destroy(Proyecto $proyecto)
     {
+        Gate::authorize('delete', $proyecto);
         try {
             $proyecto->delete();
             return response()->json(null, 204);
